Fix panic on drop.
If we dropped the whole Cursive instance while a spinner is running this caused a panic in the spinner thread. This happened because the existing drop code allowed a final frame to be set, but by this point the other end of the Cursive cb_sink has disconnected.

Fix this by refactoring the code such that a drop of Spinner avoids any further usage of the cb_sink.